requirements.txt
flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
xgboost==2.0.3
python-whois==0.8.0
dnspython==2.0.0

ipwhois==1.2.0
tld==0.13.1
tldextract==5.1.1
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
soupsieve==2.5
joblib==1.3.2
threadpoolctl==3.2.0
scipy==1.12.0 